Organize com.meti.evaluate
