+++ src/main/java/com/meti/evaluate/processable/CollectiveProcessable.java
@@ -1,5 +1,6 @@
-package com.meti.evaluate;
+package com.meti.evaluate.processable;
 
+import com.meti.evaluate.Evaluator;
 import com.meti.process.State;
 
 import java.util.Optional;
