+++ src/main/java/com/meti/evaluate/processable/MagmaPreProcessable.java
@@ -1,17 +1,19 @@
-package com.meti.evaluate;
+package com.meti.evaluate.processable;
 
 import com.meti.process.State;
 
 import java.util.function.Function;
 import java.util.stream.Stream;
 
-public class MagmaProcessable extends CollectiveProcessable {
-    public MagmaProcessable(State previous) {
+public class MagmaPreProcessable extends CollectiveProcessable {
+    public MagmaPreProcessable(State previous) {
         super(previous);
     }
 
     @Override
     protected Stream<Function<State, Processable>> streamPreprocessors() {
-        return Stream.of(VariableProcessable::new);
+        return Stream.of(
+            FunctionPreProcessable::new
+        );
     }
 }
