+++ src/main/java/com/meti/evaluate/processable/MagmaProcessable.java
@@ -1,19 +1,17 @@
-package com.meti.evaluate;
+package com.meti.evaluate.processable;
 
 import com.meti.process.State;
 
 import java.util.function.Function;
 import java.util.stream.Stream;
 
-public class MagmaPreProcessable extends CollectiveProcessable {
-    public MagmaPreProcessable(State previous) {
+public class MagmaProcessable extends CollectiveProcessable {
+    public MagmaProcessable(State previous) {
         super(previous);
     }
 
     @Override
     protected Stream<Function<State, Processable>> streamPreprocessors() {
-        return Stream.of(
-            FunctionPreProcessable::new
-        );
+        return Stream.of(VariableProcessable::new);
     }
 }
