+++ src/main/java/com/meti/evaluate/processable/Processable.java
@@ -1,5 +1,6 @@
-package com.meti.evaluate;
+package com.meti.evaluate.processable;
 
+import com.meti.evaluate.Evaluator;
 import com.meti.process.State;
 
 public interface Processable extends Evaluator<State> {
