+++ src/main/java/com/meti/evaluate/processable/VariableProcessable.java
@@ -1,4 +1,4 @@
-package com.meti.evaluate;
+package com.meti.evaluate.processable;
 
 import com.meti.stack.CallStack;
 import com.meti.render.NodeGroup;
