+++ src/main/java/com/meti/evaluate/tokenizer/CollectiveTokenizer.java
@@ -1,6 +1,7 @@
-package com.meti.evaluate;
+package com.meti.evaluate.tokenizer;
 
 import com.meti.content.Content;
+import com.meti.evaluate.Evaluator;
 import com.meti.render.Node;
 
 import java.util.Optional;
