+++ src/main/java/com/meti/evaluate/tokenizer/Tokenizer.java
@@ -1,5 +1,6 @@
-package com.meti.evaluate;
+package com.meti.evaluate.tokenizer;
 
+import com.meti.evaluate.Evaluator;
 import com.meti.render.Node;
 
 public interface Tokenizer extends Evaluator<Node> {
